Refactor: move loaders to modules
